raspberry/requirements.txt
pip>=9.0.1
python-dotenv>=0.20.0
requests>=2.25.1
setuptools
pyserial
